Refactor error handling: Introduce ErrorChain and ErrorContext modules
- Added `ErrorChain` struct for accumulating error messages with context.
- Introduced `ErrorContext` trait for adding context to results.
- Modularized error macros into separate files for better organization.
- Removed redundant error macros and consolidated them into constructors and conversions modules.
- Added tests for new error handling features and macros.
- Created a PowerShell script to analyze Rust file lengths and identify long files.
